Refactor: move `IntervalTypeOperations` tests
Moved to a base class to both be used as part of `IntervalTest`, but also for future tests independent of specific Interval types.
